+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-06-21
+
+### Changed
+
+- `mount_facet` was renamed to `bind_facet`
+- `unmount_facet` was renamed to `unbind_facet`
+- `add_component` was renamed to `bind_component`
+- `remove_component` was renamed to `unbind_component`
+- `attach` was renamed to `on_bind`
+- `detach` was renamed to `on_unbind`
+- Function signature for facet functions were normalized to use *_id params
+- Container bind/unbind methods now take `facet_id` and `component_id`
+- `DependentComponentsError` was renamed to `ComponentInUseError`
+- Exception attributes normalized to `facet_id` / `component_id`
+- GUI descriptor and method `parent` renamed to `parent_id`
+- Descriptor identity fields normalized to `id`
+
+[0.4.0]: https://github.com/Dev-DanielR/py_facetkit/releases/tag/v0.4.0
+
 ## [0.3.0] - 2026-06-20
 
 ### Added